refactor: Problem Dto 위치 이동
