+++ problem-service/src/main/java/algomarket/problemservice/application/ProblemQueryService.java
@@ -2,7 +2,7 @@ package algomarket.problemservice.application;
 
 import org.springframework.stereotype.Service;
 
-import algomarket.problemservice.application.dto.ProblemInfoResponse;
+import algomarket.problemservice.domain.problem.ProblemInfoResponse;
 import algomarket.problemservice.application.provided.ProblemFinder;
 import algomarket.problemservice.application.required.ProblemRepository;
 import algomarket.problemservice.domain.problem.Problem;
