+++ problem-service/src/main/java/algomarket/problemservice/application/provided/ProblemCreator.java
@@ -1,6 +1,6 @@
 package algomarket.problemservice.application.provided;
 
-import algomarket.problemservice.application.dto.ProblemInfoResponse;
+import algomarket.problemservice.domain.problem.ProblemInfoResponse;
 import algomarket.problemservice.domain.problem.ProblemCreateRequest;
 
 public interface ProblemCreator {
